feat: updated API error handle and refactoring in initiative invitation form
